move enum to parent package and narrow the scope of it
