+++ cms-contentful/src/test/java/com/commercetools/sunrise/cms/contentful/ContentfulCmsPageTest.java
@@ -14,7 +14,7 @@ import java.util.Locale;
 import java.util.Map;
 import java.util.Optional;
 
-import static com.commercetools.sunrise.cms.contentful.models.FieldType.*;
+import static com.commercetools.sunrise.cms.contentful.FieldType.*;
 import static java.util.Arrays.asList;
 import static java.util.Collections.singletonList;
 import static org.assertj.core.api.Assertions.assertThat;
